shiny_app/iris_social_media_downloader/requirements.txt
Django
beautifulsoup4
google-api-python-client
google-generativeai
googlemaps
matplotlib
numpy
pandas
plotly
torch
pysentimiento
python-dotenv
requests
scipy
seaborn
shiny
shinyswatch
shinywidgets
spacy
spacytextblob
transformers
tweepy
scrapy
selenium
langchain-google-genai
scrapegraphai
playwright
nest_asyncio
scraperapi-sdk
vaderSentiment
chatlas
shared
webdriver-manager
praw
deep-translator
openai
google_play_scraper
google-genai
pinecone
pinecone-client
pyvis
tabulate
smolagents
sentencepiece
supabase
psycopg2
sqlalchemy 
psycopg2-binary
selenium-stealth
uuid
openpyxl
xlsxwriter
WeasyPrint
xhtml2pdf
kaleido
lida
llmx
lida[infographics]
playwright
blinker==1.7.0
selenium-wire
geocoder
https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.8.0/es_core_news_md-3.8.0-py3-none-any.whl
